+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add filter logic to widgets for getting specific values
 * Add abstract `Partial` widget class, to easily create a new widget with a pre-defined layout
 * Wait for widgets to render their contents before starting editor
+* Add `can_be_removed` property to widget elements
 
 ## 0.1.5 (2022-07-14)